Reduce startup time
The startup time of the container was high because the image didn't contain the app.
The app was downloaded and built during startup.

I moved those stes to the Dockerfile. The advantage is faster startup. The downside
is that a new image has to be built if the app source changes, but this is considered
infrequent.